Few improvements to the versioning scheme. First changed to a global monotonically increased version number (this time really like subversion and its repo-level version number). Second returning info about what has been really deployed in remote deployment (by web service) because if you deploy foo you will end up with a package like foo-12.
git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/ode/trunk@501522 13f79535-47bb-0310-9956-ffa450edef68

+++ bpel-runtime/src/main/java/org/apache/ode/bpel/memdao/BpelDAOConnectionImpl.java
@@ -42,7 +42,7 @@ class BpelDAOConnectionImpl implements BpelDAOConnection {
         return _store.get(processId);
     }
 
-    public ProcessDAO createProcess(QName pid, QName type, String guid, int version) {
+    public ProcessDAO createProcess(QName pid, QName type, String guid, long version) {
         ProcessDaoImpl process = new ProcessDaoImpl(this,_store,pid,type, guid,version);
         _store.put(pid,process);
         return process;
