+++ bpel-store/src/main/java/org/apache/ode/store/hib/DeploymentUnitDaoImpl.java
@@ -102,7 +102,7 @@ public class DeploymentUnitDaoImpl extends HibObj implements DeploymentUnitDAO {
         super.delete();
     }
     
-    public ProcessConfDAO createProcess(QName pid, QName type, int version) {
+    public ProcessConfDAO createProcess(QName pid, QName type, long version) {
         ProcessConfDaoImpl p = new ProcessConfDaoImpl();
         p.setPID(pid);
         p.setType(type);
